hub/src/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
cffi==1.16.0
click==8.1.7
cryptography==43.0.0
html5tagger==1.3.0
httptools==0.6.1
multidict==6.0.5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.1
sanic==24.6.0
sanic-ext==23.12.0
sanic-routing==23.12.0
setuptools==71.1.0
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
uvloop==0.19.0
websockets==12.0

aiohttp==3.8.5
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.0
uvicorn[standard]==0.22.0
httpx==0.24.1